+++ packages/adhere-util-resource/lib/dict/dict.moment.config.d.ts
@@ -1,5 +1,4 @@
 declare namespace _default {
-    export function initStatic(): void;
-    export function initStatic(): void;
+    function initStatic(): void;
 }
 export default _default;